+++ test/core/do_while.sv
@@ -1,0 +1,11 @@
+module top;
+    integer x = 0;
+    initial
+        do begin
+            $display("hi %0d", x);
+            x++;
+            if (x == 2)
+                continue;
+            $display("step");
+        end while (0 < x && x < 3);
+endmodule