+++ test/nosim/do_while.sv
@@ -1,5 +1,0 @@
-module top;
-    initial do
-        $display("hi");
-    while (0);
-endmodule