+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added support for enumerated type ranges (e.g., `enum { X[3:5] }`)
 * Added support for the SystemVerilog `edge` event
 * Added support for cycle delay ranges in assertion sequence expressions
+* Added conversion for `do` `while` loops
 * Added support for passing through DPI imports and exports
 * Added support for passing through functions with output ports